New ShortWalkView, which instructs the user on the use of the timed walk. Includes a plist for content.
Add padding to the GenericInstructionView_Previews layout

When these were added, codesign complained that DASIQuestions.json had not been "signed at all". Failed to help:
Deleting the Derived Data for the project and restarting Xcode
Updating certificates, verifying the validity of the (Development) provisioning profile.
Removing DASIQuestions.json from the Async Accel build order resolved the error, but obviously we need that file.
Trying a build of Better Step with/without the file is impractical, because that target is abandoned and the code no longer compiles.