Have more types adopt CSVRepresentable.
CMAcceleration, CMAccelerometerDate, AnswerState, DASIResponseList, DASIUserResponse, TimedXYZRepresentable
New: csvData, which doesn't seem to be accepted for [String] (I think)

A few #warning reminders, including that

Add AppStorage key to flag completion of the 7-day report; also new daysSince7DayReport  to handle overlaps and gaps in the 7-day periods.
Some whitespace

This is the first commit from a large number of changes across many files. The overall goal is to push data up from the tasks to the PhaseStorage instance.