This build demonstrates the full end-to-end workflow of Better Step.
It looks even better from the inside: All phases (but the 8-day step count) generate apparently correct data. However, something had to give in this release: No archiving is done, and there is no human access to the (nonetheless correct) .csv files.

The gear button is stil in the top-right corner. Subjects will never see it, but for testing, you'd want to restore the app to ground zero without having to reinstall it. HOWEVER, this is erratic, and I can't guarantee anything useful would come of rapping it. This is a leading priority.   — F