Remove the commentary on PedometryFromHealthKit/stiore.
Remove "import stupload" from TopContainerView, PerformUpload, UploadTaskDelegate
Not going to accede to HKHealthStore() being an optional return. , when you treat fhe init as unwrapped, it demands optional, and vice versa.
Attempt DateFormatter()! for the formatters in error-support.swift. (Again, when you treat fhe init as unwrapped, it demands optional, and vice versa.)
Playground code to try figuring out the is/isn't a failable initializer.

Changes to the .pbxproj will come in the next commit or two.