Better Step 1.0 (31?) 21-Feb-2023
Restores spoken audio for walking exercises.
Restores "< Back" button to interstitial page series.

This version has not yet made the "loudness" image a first-class part of that series. it does not include a "< Back" button.

(internal) Improved response to and handling of occasions to clear out the collected data. The intended use is to reset the app for a fresh session. I’ve remarked before on the difficulty of getting this right.

Error handling is inadequate. There are two occasions for error in this app: (1) internal - I made a coding error, or the network is unavailable, or the server doesn't respond or appears not to exist. (2) server does receives the request but does/can not handle it (resource not found, authentication required, general server error, I'm a teapot…). https://en.wikipedia.org/wiki/List_of_HTTP_status_codes

A well-behaved app would post an alert explaining the problem as best it can, and offering a retry. That's now what happens here: If there's any error for any reason, Better Step throws up its hands, discards the collected data, and rewinds to the start of the workflow.