First cut at submission only at the end of the Conclusion page, not at submission of all required phases.
This is a bit tricky, because submission had been hard-wired into PhaseStorage as soon as all phases had presented their CSV content.
This was very robust and worked well. Now the app relies on correct a ReportingPhase callback from ConclusionView,

If a second page were added to ConclusionView,
the callback would have to be at the end of the _FIRST,_ not at the end of the phase.
BUT there's  a second state to ConclusionView, the release from the last card and back to the first phase.