Signing/metadata error: com.drmobile.Walk-Test in bundle ≠ "com.drdrlabs.walk-test in the App Store listing.
Issue a DEVELOPMENT identity.
Decide not to request health-records privilege.  That's a HealthKit privilege, and I don't use HK that way(?)
Change the iin-project/Info.plist  to use com.drdrmobile.walk-test. This isn't desirable (won't be a reserved host name) but this is easier than registering a new app + ID.
Remove the healthkit